requirements.txt
rich>=13.0.0
pyyaml>=6.0
requests>=2.31.0
jinja2>=3.1.0
weasyprint>=60.0
svglib>=1.5.0
reportlab>=4.0.0
